Major app update.
The login and signup now works 90%. you can login if you've signed up already. You can sign up and all your details will be added to the database. keep it mind that all the data is not stored currently, but only the neccessary one. Everyone please sign up so we can capture the data to the database.